Add functionality to load charts by pure string
The FNFSong class has been modified to permit construction of itself using a string on its own that is passed into the constructor, as long as you tell it to load it "AsRawJson".

This is useful when you can't have a compiled application reading external files but still want to have to construct a FNFSong class.